asciiio: utilize high resolution monotonic clock when available
time.time()'s resolution (despite its precision) is documented to be (at worst)
1-second granularity, which is terrible for something like screencasting:

http://docs.python.org/2/library/time.html#time.time

The clocks time.time() tries are (in order):

	- gettimeofday()
	- ftime()
	- time()

None of these clocks are monotonically increasing, and the highest resolution
of these is gettimeofday, which is (at least on my system) 1MHz.

If we were using Python 3.x, we could just use time.monotonic() and be done
with it, but we're in a largely Python 2.x world right now. For now, let's use
ctypes to load up and use clock_gettime.

clock_gettime(CLOCK_MONOTONIC,...) exemplifies all the attributes we care about
for this particular use case: it's monotonically increasing, it's high
precision, and it's high resolution. This makes it ideal for determining
timespans.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>