Add -t/--time option to play command
option -t / --time allows user to start replay of a file at a
specified sec. It's quite stupid implemetnation, but should be enough for
review of a big file.

The logic behind is in Player class where sum of delays runs first in
every loop and if this sum is bigger than intended time, than we start
replaying.

This is a part of solution of Issue #118.