Check playback size can fit in terminal
Before playing a recorded file, check that the terminal that is running
the playback is at least as large as the recorded content.  This will
allow all the escape characters that change the location on the screen
to go to a legal spot and, therefore, everything should display
properly.

I've chosen to warn for a smaller size instead of causing an error in
case the playback doesn't use escape codes that are too large, but one
could argue that an error is legitimate here, too.