Consume terminal answerbacks in asciinema cat
If the asciicast contains escape sequence queries like "CSI 6 n", the
terminal will both echo and input its responses.  Use the raw() context
manager that we already have to attempt to consume this "input."

There is, unfortunately, no way of finding out exactly when the terminal
has finished its answerbacks.  This patch adds a 50ms wait, which should
be overkill for a local terminal (in my tests, 3ms was usually enough).
For a remote terminal, this number becomes harder to estimate.

Technically this only needs to be done if whatever we're writing to
isatty(), but keep it simple for now.